+++ src/partials/html.js
@@ -2,8 +2,6 @@ module.exports = [
     {
         "scope": [
             "punctuation.definition.tag",
-            "meta.tag.sgml.doctype.html",
-            "meta.tag.sgml.html",
         ],
         "settings": { "foreground": "#808080" }
     },
@@ -24,8 +22,6 @@ module.exports = [
             "punctuation.section.embedded.end.php",
             "meta.tag.template.value.twig",
             "punctuation.section.tag.twig",
-            "punctuation.definition.generic.begin.html",
-            "punctuation.definition.generic.end.html",
         ],
         "settings": { "foreground": "#569cd6" }
     },
@@ -34,24 +30,29 @@ module.exports = [
         "settings": { "foreground": "#dcdcaa" }
     },
     {
-        "name": "ejs and erb",
+        "name": "don't highlight ejs and erb as errors",
         "scope": "invalid.illegal.bad-angle-bracket.html",
         "settings": { "foreground": "#d4d4d4" }
     },
     {
-        "scope": "variable.parameter.handlebars",
+        "scope": [
+            "variable.parameter.handlebars",
+        ],
         "settings": { "foreground": "#9cdcfe" }
     },
     {
+        "name": "Mustaches",
         "scope": [
-            "meta.directive.vue",
-            "expression.embedded.vue",
-            "support.constant.handlebars",
+            "punctuation.definition.generic.begin.html",
+            "punctuation.definition.generic.end.html",
             "meta.function.inline.other.handlebars",
-            "meta.first-class.jade",
-            "meta.control.flow.jade",
-            "storage.type.function.jade",
+            "support.constant.handlebars",
+            "expression.embedded.vue",
+            "meta.directive.vue",
             "storage.type.import.include.jade",
+            "storage.type.function.jade",
+            "meta.control.flow.jade",
+            "meta.first-class.jade",
         ],
         "settings": { "foreground": "#c586c0" }
     },
@@ -62,31 +63,4 @@ module.exports = [
         ],
         "settings": { "foreground": "#ce9178" }
     },
-    // {
-    //     "scope": "meta.embedded.line.erb",
-    //     "settings": {
-    //         "foreground": "#9CDCFE"
-    //     }
-    // },
-    // {
-    //     "scope": [
-    //         "punctuation.section.embedded.begin.erb",
-    //         "punctuation.section.embedded.end.erb"
-    //     ],
-    //     "settings": {
-    //         "foreground": "#808080"
-    //     }
-    // },
-    // {
-    //     "scope": [
-    //         "source.ruby",
-    //         "text.html.erb",
-    //         "text.haml",
-    //         "punctuation.separator.object.ruby",
-    //         "punctuation.separator.other.ruby"
-    //     ],
-    //     "settings": {
-    //         "foreground": "#d4d4d4"
-    //     }
-    // },
 ]